+++ velox/functions/sparksql/CMakeLists.txt
@@ -27,6 +27,7 @@ add_subdirectory(specialforms)
 velox_add_library(
   velox_functions_spark_impl
   ArrayGetFunction.cpp
+  BloomFilter.cpp
   ArraySort.cpp
   AssertNotNull.cpp
   SequenceFunction.cpp
@@ -54,6 +55,7 @@ velox_add_library(
   XxHash64.cpp
   HEADERS
   AnsiMode.h
+  BloomFilter.h
   Arena.h
   Arithmetic.h
   ArrayAppend.h
