+++ velox/functions/sparksql/aggregates/BloomFilterAggAggregate.cpp
@@ -16,7 +16,7 @@
 
 #include "velox/functions/sparksql/aggregates/BloomFilterAggAggregate.h"
 
-#include "velox/common/base/BloomFilter.h"
+#include "velox/functions/sparksql/BloomFilter.h"
 #include "velox/exec/Aggregate.h"
 #include "velox/expression/FunctionSignature.h"
 #include "velox/functions/sparksql/SparkQueryConfig.h"
