+++ velox/functions/sparksql/tests/MightContainTest.cpp
@@ -15,7 +15,7 @@
  */
 
 #include "velox/functions/sparksql/MightContain.h"
-#include "velox/common/base/BloomFilter.h"
+#include "velox/functions/sparksql/BloomFilter.h"
 #include "velox/core/Expressions.h"
 #include "velox/functions/sparksql/tests/SparkFunctionBaseTest.h"
